Fix macro paths
The macro's were already `#[macro_export]`ed, but this (a) makes them
available on the root of the crate rather than in the module they are in
and (b) only does so _outside_ the crate, making it impossible to use a
recursive macro inside the crate.

By adding a re-export (`use macro_name as some_other_name`) the macro
can be re-exported as a member of the module it is in, solving both of
these issues.

Note that by defualt re-exporting a macro like this can only be done
with `pub(crate)` visibility, it _must_ be paired with the
`#[macro_export]` to allow `pub`.