Optimize day05 part 2
This is done by tracking and transforming ranges instead of single seeds
numbers.

This could also be used to solve part 1 by just creating a bunch of
ranges of length 1, but the added complexity of range processing makes
this about 3 times slower than the simple implementation, so the old
solution is kept for part 1.