Replace threadpool with rayon
The main motivation for this is that rayon has WASM compatibility,
whereas threadpool does not. It'd be possible to create something to
make it use an alternative thread(pool) library on WASM like we do for
timing, but this is easier.

An added benefit is that rayon already provides functions on iterators
to parallelize meaning we no longer have to implement threaded_*
functions ourselve

It also seems to perform slightly better.