Optimize 21-2
I was taking steps of 2*bounds when collecting the datapoints for the
interpolation due to a bug I had early on where the naive method yielded
incorrect results for even step counds, and I never changed this even
after fixing this bug.

By adjusting this to steps of 1*bounds the amount of steps that need to
be calculated by the naive method drops from 589 to 327, which results
in a substantial performance improvement.